fix: hide Ellipsis component
