+++ packages/picasso-lab/src/TypographyOverflow/TypographyOverflow.tsx
@@ -1,7 +1,8 @@
 import React, { ReactNode, ReactElement } from 'react'
 import { BaseProps } from '@toptal/picasso-shared'
-import { Ellipsis } from '@toptal/picasso-lab'
 import { Tooltip } from '@toptal/picasso'
+
+import Ellipsis from '../Ellipsis'
 
 export interface Props extends BaseProps {
   /** A typography which can possibly overflow */
