+++ packages/picasso-lab/src/TypographyOverflow/story/Default.example.tsx
