+++ packages/picasso-lab/src/index.ts
@@ -4,5 +4,4 @@ export { default as OverviewBlock } from './OverviewBlock'
 export { default as Subheader } from './Subheader'
 export { default as Breadcrumbs } from './Breadcrumbs'
 export { default as List } from './List'
-export { default as Ellipsis } from './Ellipsis'
 export { default as TypographyOverflow } from './TypographyOverflow'
